feat(rust): Move setters and getters of options to the adbc_ffi crate
