+++ rust/ffi/src/lib.rs
@@ -39,7 +39,7 @@
 //! [export_driver] macro.
 
 pub mod driver_exporter;
-mod options;
+pub mod options;
 #[doc(hidden)]
 pub use driver_exporter::FFIDriver;
 pub mod methods;
